@@ -18,7 +18,7 @@ description: "A new Flutter module project."
 version: 1.0.0+1
 
 environment:
-  sdk: '3.2.4'
+  sdk: '3.3.4'
 
 dependencies:
   flutter:
@@ -42,6 +42,8 @@ dependencies:
   hive_flutter: ^1.1.0
   path_provider: ^2.1.3
   jwt_decode: ^0.3.1
+  contacts_service: ^0.6.3
+  permission_handler: ^11.3.1
 
 dev_dependencies:
   flutter_test:
