@@ -1,19 +1,14 @@
 
-import 'package:call_log/call_log.dart';
-import 'package:hive/hive.dart';
+import 'package:secucalls/constant/constants.dart';
+// Unique typeId for each Hive model
+class MyNumberCallLog {
 
-part 'my_number_call_log_model.g.dart';
-
-@HiveType(typeId: 0) // Unique typeId for each Hive model
-class MyNumberCallLog extends HiveObject {
-  @HiveField(0)
   final String cachedMatchedNumber;
 
-  @HiveField(1)
-  final CallType callType;
+  final TypeOfCall callType;
 
-  @HiveField(2)
-  final CallType name;
+
+  final String name;
 
   MyNumberCallLog(this.cachedMatchedNumber, this.callType, this.name);
 }
